A refactoring effort involving extracting facades out of utils, fixing singleton and more
feat: Implement RoomViewModel for room searching functionality

feat: Create IndoorNavigationFacade for managing indoor navigation state and actions

feat: Develop MapFacade for handling map-related operations and state management

feat: Introduce POIFacade for managing Points of Interest (POI) operations

feat: Add ShuttleFacade for real-time shuttle data fetching and route calculations
