+++ src/MVVM/repositories/BuildingRepository.ts
@@ -1,11 +1,26 @@
 import { Building } from "@/MVVM/models/Building";
 import { IBuildingRepository } from './Interfaces/IBuildingRepository';
-import { BaseRepository } from "./BaseRepository";
 
 // Export the interface for backward compatibility
 export { IBuildingRepository as BuildingRepository };
 
-export class BuildingRepositoryImpl extends BaseRepository<Building> implements IBuildingRepository {
+export class BuildingRepositoryImpl implements IBuildingRepository {
+  // The single instance
+  private static instance: BuildingRepositoryImpl | null = null;
+  
+  // Private constructor ensures singleton pattern
+  private constructor() {
+    // Initialize any resources needed
+  }
+
+  // Public static method to get the singleton instance
+  public static getInstance(): BuildingRepositoryImpl {
+    if (!BuildingRepositoryImpl.instance) {
+      BuildingRepositoryImpl.instance = new BuildingRepositoryImpl();
+    }
+    return BuildingRepositoryImpl.instance;
+  }
+
   async findBuildingById(id: string): Promise<Building> {
     throw new Error("Method not implemented: findBuildingById");
   }
