+++ src/MVVM/repositories/CalendarRepository.ts
@@ -1,10 +1,25 @@
 import { ICalendarRepository } from "./Interfaces/ICalendarRepository";
-import { BaseRepository } from "./BaseRepository";
 
 // Export the interface for backward compatibility
 export { ICalendarRepository as CalendarRepository };
 
-export class CalendarRepositoryImpl extends BaseRepository<any> implements ICalendarRepository {
+export class CalendarRepositoryImpl implements ICalendarRepository {
+  // The single instance
+  private static instance: CalendarRepositoryImpl | null = null;
+  
+  // Private constructor ensures singleton pattern
+  private constructor() {
+    // Initialize any resources needed
+  }
+
+  // Public static method to get the singleton instance
+  public static getInstance(): CalendarRepositoryImpl {
+    if (!CalendarRepositoryImpl.instance) {
+      CalendarRepositoryImpl.instance = new CalendarRepositoryImpl();
+    }
+    return CalendarRepositoryImpl.instance;
+  }
+
   async connectGoogleCalendar(googleCalendarId: string, userId: string): Promise<any> {
     throw new Error("Method not implemented: connectGoogleCalendar");
   }
