+++ src/MVVM/repositories/CampusRepository.ts
@@ -1,11 +1,26 @@
 import { Campus } from '@/MVVM/models/Campus';
 import { ICampusRepository } from './Interfaces/ICampusRepository';
-import { BaseRepository } from './BaseRepository';
 
 // Export the interface for backward compatibility
 export { ICampusRepository as CampusRepository };
 
-export class CampusRepositoryImpl extends BaseRepository<Campus> implements ICampusRepository {
+export class CampusRepositoryImpl implements ICampusRepository {
+  // The single instance
+  private static instance: CampusRepositoryImpl | null = null;
+  
+  // Private constructor ensures singleton pattern
+  private constructor() {
+    // Initialize any resources needed
+  }
+
+  // Public static method to get the singleton instance
+  public static getInstance(): CampusRepositoryImpl {
+    if (!CampusRepositoryImpl.instance) {
+      CampusRepositoryImpl.instance = new CampusRepositoryImpl();
+    }
+    return CampusRepositoryImpl.instance;
+  }
+
   /**
    * Retrieves a campus by its unique identifier
    * @param id - The string of the campus to find
