+++ src/MVVM/repositories/RoomRepository.ts
@@ -1,17 +1,23 @@
 import { Room, Building } from '../models/Room';
 import * as fs from 'fs';
 import * as path from 'path';
-import { getErrorMessage } from '@/Shared/utils/generalUtils';
-import { BaseRepository } from './BaseRepository';
+import { getErrorMessage } from '@/Shared/utils/GeneralUtils';
 import { IRoomRepository } from './Interfaces/IRoomRepository';
 
-export class RoomRepository extends BaseRepository<Building> implements IRoomRepository {
+export class RoomRepositoryImpl implements IRoomRepository {
   private readonly dataDir: string;
+  private static instance: RoomRepositoryImpl | null = null;
 
-  protected constructor(dataDir: string = path.join(process.cwd(), 'data')) {
-    super();
+  private constructor(dataDir: string = path.join(process.cwd(), 'data')) {
     this.dataDir = dataDir;
     this.ensureDataDirectory();
+  }
+
+  public static getInstance(): RoomRepositoryImpl {
+    if (!RoomRepositoryImpl.instance) {
+      RoomRepositoryImpl.instance = new RoomRepositoryImpl();
+    }
+    return RoomRepositoryImpl.instance;
   }
 
   private ensureDataDirectory(): void {
@@ -103,3 +109,6 @@ export class RoomRepository extends BaseRepository<Building> implements IRoomRep
     }
   }
 }
+
+// Export the interface for backward compatibility
+export { IRoomRepository as RoomRepository };
