+++ src/MVVM/repositories/RouteRepository.ts
@@ -1,46 +1,61 @@
 import { Location } from "@/MVVM/models/Location";
 import { Route, RouteSegment, TransportationMode } from "@/MVVM/models/Route";
 import { IRouteRepository } from './Interfaces/IRouteRepository';
-import { BaseRepository } from "./BaseRepository";
 
 // Export the interface for backward compatibility
 export { IRouteRepository as RouteRepository };
 
-export class RouteRepositoryImpl extends BaseRepository<Route> implements IRouteRepository {
-    // Read operations
-    async findRouteById(id: string): Promise<Route> {
-        throw new Error("Method not implemented: findRouteById");
-    }
+export class RouteRepositoryImpl implements IRouteRepository {
+  // The single instance
+  private static instance: RouteRepositoryImpl | null = null;
+  
+  // Private constructor ensures singleton pattern
+  private constructor() {
+    // Initialize any resources needed
+  }
 
-    async findSegmentsByRouteId(routeId: string): Promise<RouteSegment[]> {
-        throw new Error("Method not implemented: findSegmentsByRouteId");
+  // Public static method to get the singleton instance
+  public static getInstance(): RouteRepositoryImpl {
+    if (!RouteRepositoryImpl.instance) {
+      RouteRepositoryImpl.instance = new RouteRepositoryImpl();
     }
+    return RouteRepositoryImpl.instance;
+  }
 
-    // Create operations
-    async createRoute(waypoints: Location[], mode: TransportationMode, userId: string): Promise<Route> {
-        throw new Error("Method not implemented: createRoute");
-    }
+  // Read operations
+  async findRouteById(id: string): Promise<Route> {
+    throw new Error("Method not implemented: findRouteById");
+  }
 
-    async findOrCreateSegment(startPoint: Location, endPoint: Location, mode: TransportationMode, userId: string): Promise<RouteSegment> {
-        throw new Error("Method not implemented: findOrCreateSegment");
-    }
+  async findSegmentsByRouteId(routeId: string): Promise<RouteSegment[]> {
+    throw new Error("Method not implemented: findSegmentsByRouteId");
+  }
 
-    // Update operations
-    async updateRoute(id: string, updates: Partial<Route>, userId: string): Promise<Route> {
-        throw new Error("Method not implemented: updateRoute");
-    }
+  // Create operations
+  async createRoute(waypoints: Location[], mode: TransportationMode, userId: string): Promise<Route> {
+    throw new Error("Method not implemented: createRoute");
+  }
 
-    async updateSegment(segmentId: string, updates: Partial<RouteSegment>, userId: string): Promise<RouteSegment> {
-        throw new Error("Method not implemented: updateSegment");
-    }
+  async findOrCreateSegment(startPoint: Location, endPoint: Location, mode: TransportationMode, userId: string): Promise<RouteSegment> {
+    throw new Error("Method not implemented: findOrCreateSegment");
+  }
 
-    // Delete operations
-    async deleteRoute(id: string, userId: string): Promise<void> {
-        throw new Error("Method not implemented: deleteRoute");
-    }
+  // Update operations
+  async updateRoute(id: string, updates: Partial<Route>, userId: string): Promise<Route> {
+    throw new Error("Method not implemented: updateRoute");
+  }
 
-    // Calculation
-    async calculatePath(segmentId: string, userId: string): Promise<RouteSegment> {
-        throw new Error("Method not implemented: calculatePath");
-    }
+  async updateSegment(segmentId: string, updates: Partial<RouteSegment>, userId: string): Promise<RouteSegment> {
+    throw new Error("Method not implemented: updateSegment");
+  }
+
+  // Delete operations
+  async deleteRoute(id: string, userId: string): Promise<void> {
+    throw new Error("Method not implemented: deleteRoute");
+  }
+
+  // Calculation
+  async calculatePath(segmentId: string, userId: string): Promise<RouteSegment> {
+    throw new Error("Method not implemented: calculatePath");
+  }
 }
