+++ src/MVVM/repositories/UserRepository.ts
@@ -1,11 +1,26 @@
 import { User } from "@/MVVM/models/User";
 import { IUserRepository } from "./Interfaces/IUserRepository";
-import { BaseRepository } from "./BaseRepository";
 
 // Export the interface for backward compatibility
 export { IUserRepository as UserRepository };
 
-export class UserRepositoryImpl extends BaseRepository<User> implements IUserRepository {
+export class UserRepositoryImpl implements IUserRepository {
+  // The single instance
+  private static instance: UserRepositoryImpl | null = null;
+  
+  // Private constructor ensures singleton pattern
+  private constructor() {
+    // Initialize any resources needed
+  }
+
+  // Public static method to get the singleton instance
+  public static getInstance(): UserRepositoryImpl {
+    if (!UserRepositoryImpl.instance) {
+      UserRepositoryImpl.instance = new UserRepositoryImpl();
+    }
+    return UserRepositoryImpl.instance;
+  }
+
   async getUserById(userId: string): Promise<User> {
     throw new Error("Method not implemented: getUserById");
   }
